- Added UI And Reducer of estimation Feature edit by Estimator
