+++ client/components/sections/EstimationFeatures.js
@@ -31,8 +31,12 @@ class EstimationFeature extends React.PureComponent {
                 if (feature.addedInThisIteration) {
                     logger.debug(logger.ESTIMATION_FEATURE_BUTTONS, 'added in this iteration, edit button')
                     // Estimator would see plain edit button in case he has added task in this iteration
-                    buttons.push(<img key="edit" src="/images/edit.png"></img>)
-                    buttons.push(<img key="edit" src="/images/delete.png"></img>)
+                    buttons.push(<img key="edit" src="/images/edit.png" onClick={() => {
+                        this.props.showEditFeatureForm(feature)
+                    }}></img>)
+                    buttons.push(<img key="delete" src="/images/delete.png" onClick={() => {
+                        this.props.deleteFeature(feature)
+                    }}></img>)
                 } else {
                     if (feature.negotiator.changeRequested) {
                         logger.debug(logger.ESTIMATION_FEATURE_BUTTONS, 'negotiator requested change, he_requested_edit button')
@@ -99,6 +103,8 @@ class EstimationFeature extends React.PureComponent {
 let
     EstimationFeatures = (props) =>
         Array.isArray(props.features) && props.features.map(f => <EstimationFeature feature={f} key={f._id}
+                                                                                    showEditFeatureForm={props.showEditFeatureForm}
+                                                                                    deleteFeature={props.deleteFeature}
                                                                                     loggedInUserRole={props.loggedInUserRole}/>)
 
 export default EstimationFeatures
